feat: add search package utils
