+++ .changeset/six-boats-sparkle.md
@@ -1,0 +1,5 @@
+---
+'@verdaccio/search': minor
+---
+
+feat: add search package utilities
